Increase contrast of color output
On some color schemes (like the one I use), the text could to read when
white and black are more muted. For example, if white is closer to a
light gray. This sets the foreground to either RGB `#FFFFFF` or
`#000000`, so that there should always be a reasonable amount of
contrast between the colorized background and the foreground text.